requirements.txt
certifi==2021.10.8
mysql-connector-python==8.0.27
numpy==1.21.4
pandas==1.3.4
protobuf==3.19.1
pyodbc==4.0.32
pyproj==3.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
Shapely==1.8.0
six==1.16.0
sql==0.4.0
